Add React and Vue jsx runtime support
