+++ packages/xoid/copy/jsx-runtime.js
@@ -1,0 +1,1 @@
+module.exports = {}
