+++ packages/xoid/src/jsx-runtime.d.ts
@@ -1,0 +1,1 @@
+type A = JSX.Element
