+++ packages/xoid/src/jsx-runtime.js
@@ -1,0 +1,1 @@
+module.exports = {}
