--- packages/xoid/src/jsx-runtime.tsx
@@ -1,1 +1,0 @@
-export * from 'react/jsx-runtime'
