+++ tests/isomorphism/dependency-injection.test.tsx
@@ -1,6 +1,8 @@
 import { Adapter, InjectionKey } from 'xoid'
-import toReact, { useSetup as useSetupReact } from '@xoid/react'
-import toVue, { useSetup as useSetupVue } from '@xoid/vue'
+import { useSetup as useSetupReact } from '@xoid/react'
+import toReact from '@xoid/react/src/runtime'
+import { useSetup as useSetupVue } from '@xoid/vue'
+import toVue from '@xoid/vue/src/runtime'
 import { render as renderReact } from '@testing-library/react'
 import { render as renderVue } from '@testing-library/vue'
 import React from 'react'
