+++ tests/isomorphism/setup-and-markup/CounterReact.tsx
@@ -1,4 +1,4 @@
-import toReact from '@xoid/react'
+import toReact from '@xoid/react/runtime'
 import Counter from './Counter'
 
 const CounterReact = toReact(Counter)
