+++ tests/isomorphism/setup-and-markup/CounterVue.tsx
@@ -1,4 +1,4 @@
-import toVue from '@xoid/vue'
+import toVue from '@xoid/vue/runtime'
 import Counter from './Counter'
 
 const CounterVue = toVue(Counter)
