+++ tsconfig.json
@@ -4,6 +4,9 @@
     "baseUrl": ".",
     "paths": {
       "xoid": ["./packages/xoid/src"],
+      "xoid/jsx-runtime": ["./packages/xoid/src/jsx-runtime"],
+      "@xoid/react/runtime": ["./packages/react/src/runtime"],
+      "@xoid/vue/runtime": ["./packages/vue/src/runtime"],
       "@xoid/*": ["./packages/*/src"]
     },
     "target": "es5",
